% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa-functions.R
\name{calc_null_effect}
\alias{calc_null_effect}
\title{Function to calculate null-effect sequence ratios}
\usage{
calc_null_effect(event_times, weight_dates, obs_period, window_size = 365)
}
\description{
Function to calculate null-effect sequence ratios
}
