% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_fxns.R
\name{expit}
\alias{expit}
\title{Expit function}
\usage{
expit(x)
}
\arguments{
\item{x}{value between -infinity and infinity}
}
\value{
1/(1+exp(-x))
}
\description{
Expit function
}
\examples{
x <- 1;
expit(x);

}
