% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa-functions.R
\name{sr_histogram}
\alias{sr_histogram}
\title{Function to produce a barchart representing the distribution of sequences
over time}
\usage{
sr_histogram(t, binsize = 30)
}
\description{
Function to produce a barchart representing the distribution of sequences
over time
}
