% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa-functions.R
\name{sr_wrapper}
\alias{sr_wrapper}
\title{Function to obtain a crude sequence ratio from exposure and event times
for each individual}
\usage{
sr_wrapper(
  id_expo,
  t_expo,
  id_event,
  t_event,
  window_size = 365,
  blackout = 0,
  obs_period = NULL,
  weights = "studypop"
)
}
\description{
Function to obtain a crude sequence ratio from exposure and event times
for each individual
}
