% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\title{ssa}
\usage{
ssa(a, b, d, e = 1)
}
\arguments{
\item{a}{path to docx template file}

\item{b}{the first number of the row indicating the first minute of observation}

\item{d}{the last number of the row indicating the last minute of observation}

\item{e}{a number indicating the minimum number for transitions between sequences}
}
\description{
ssa
}
